Refactoring GUI out of the ReportManager.
This will let us later use the GUI code for the Trashcan.
